custom_llm/requirements.txt
Package                       Version
----------------------------- ------------
accelerate                    0.21.0
adapter-transformers          3.2.1.post0
aiohttp                       3.8.6
aiosignal                     1.3.1
altair                        5.1.2
anyio                         3.7.1
asttokens                     2.4.1
async-timeout                 4.0.3
attrs                         23.1.0
awswrangler                   3.4.2
backports.functools-lru-cache 1.6.5
beautifulsoup4                4.12.2
bitsandbytes                  0.40.2
blinker                       1.7.0
boto3                         1.33.8
botocore                      1.33.8
bs4                           0.0.1
cachetools                    5.3.2
certifi                       2023.7.22
charset-normalizer            3.3.2
click                         8.1.7
comm                          0.1.4
dataclasses-json              0.6.2
datasets                      2.14.7
debugpy                       1.6.7
decorator                     5.1.1
dill                          0.3.7
einops                        0.7.0
exceptiongroup                1.1.3
executing                     2.0.1
fastapi                       0.104.1
filelock                      3.13.1
frozenlist                    1.4.0
fsspec                        2023.10.0
gcloud                        0.18.3
gitdb                         4.0.11
GitPython                     3.1.40
google                        3.0.0
google-api-core               2.14.0
google-auth                   2.25.1
google-cloud-bigquery         3.13.0
google-cloud-core             2.3.3
google-cloud-storage          2.14.0
google-crc32c                 1.5.0
google-resumable-media        2.6.0
googleapis-common-protos      1.61.0
greenlet                      3.0.1
grpcio                        1.59.3
grpcio-status                 1.59.3
h11                           0.14.0
httplib2                      0.22.0
httptools                     0.6.1
huggingface-hub               0.20.2
idna                          3.4
importlib-metadata            6.8.0
ipykernel                     6.26.0
ipython                       8.17.2
iso-639                       0.4.5
iso639                        0.1.4
iso639-lang                   2.2.0
jedi                          0.19.1
Jinja2                        3.1.2
jmespath                      1.0.1
jsonpatch                     1.33
jsonpointer                   2.4
jsonschema                    4.19.2
jsonschema-specifications     2023.7.1
jupyter_client                8.6.0
jupyter_core                  5.5.0
langchain                     0.0.339
langsmith                     0.0.66
markdown-it-py                3.0.0
MarkupSafe                    2.1.3
marshmallow                   3.20.1
matplotlib-inline             0.1.6
mdurl                         0.1.2
mpmath                        1.3.0
msgpack                       1.0.7
multidict                     6.0.4
multiprocess                  0.70.15
mypy-extensions               1.0.0
mysql-connector-python        8.2.0
nest-asyncio                  1.5.8
networkx                      3.2.1
ninja                         1.11.1.1
numpy                         1.26.2
nvidia-cublas-cu12            12.1.3.1
nvidia-cuda-cupti-cu12        12.1.105
nvidia-cuda-nvrtc-cu12        12.1.105
nvidia-cuda-runtime-cu12      12.1.105
nvidia-cudnn-cu12             8.9.2.26
nvidia-cufft-cu12             11.0.2.54
nvidia-curand-cu12            10.3.2.106
nvidia-cusolver-cu12          11.4.5.107
nvidia-cusparse-cu12          12.1.0.106
nvidia-nccl-cu12              2.18.1
nvidia-nvjitlink-cu12         12.3.52
nvidia-nvtx-cu12              12.1.105
oauth2client                  4.1.3
packaging                     23.2
pandas                        2.1.3
parso                         0.8.3
peft                          0.4.0
pexpect                       4.8.0
pickleshare                   0.7.5
Pillow                        10.1.0
pip                           23.3
platformdirs                  4.0.0
prompt-toolkit                3.0.41
proto-plus                    1.22.3
protobuf                      4.21.12
psutil                        5.9.5
psycopg2-binary               2.9.9
ptyprocess                    0.7.0
pure-eval                     0.2.2
pyarrow                       14.0.1
pyarrow-hotfix                0.5
pyasn1                        0.5.1
pyasn1-modules                0.3.0
pydantic                      1.10.13
pydeck                        0.8.1b0
Pygments                      2.16.1
pyparsing                     3.1.1
python-dateutil               2.8.2
python-dotenv                 1.0.0
pytz                          2023.3.post1
PyYAML                        6.0.1
pyzmq                         23.0.0
ray                           2.8.0
referencing                   0.30.2
regex                         2023.10.3
requests                      2.31.0
rich                          13.6.0
rpds-py                       0.12.0
rsa                           4.9
s3transfer                    0.8.2
safetensors                   0.4.0
scipy                         1.11.3
sentencepiece                 0.1.99
setuptools                    68.0.0
six                           1.16.0
smmap                         5.0.1
sniffio                       1.3.0
soupsieve                     2.5
SQLAlchemy                    2.0.23
stack-data                    0.6.2
starlette                     0.27.0
streamlit                     1.28.2
streamlit-sync                0.1.0
sympy                         1.12
tenacity                      8.2.3
tiktoken                      0.5.1
tokenizers                    0.13.0
toml                          0.10.2
toolz                         0.12.0
torch                         2.1.0
tornado                       6.3.3
tqdm                          4.66.1
traitlets                     5.13.0
transformers                  4.31.0
triton                        2.1.0
trl                           0.4.7
typing_extensions             4.8.0
typing-inspect                0.9.0
tzdata                        2023.3
tzlocal                       5.2
urllib3                       1.26.18
uvicorn                       0.24.0.post1
uvloop                        0.19.0
validators                    0.22.0
vllm                          0.2.2
watchdog                      3.0.0
watchfiles                    0.21.0
wcwidth                       0.2.10
websockets                    12.0
wheel                         0.41.2
xformers                      0.0.22.post7
xxhash                        3.4.1
yarl                          1.9.2
zipp                          3.17.0
